+++ choco-solver/src/main/java/solver/constraints/extension/nary/LargeRelation.java
@@ -27,6 +27,8 @@
 
 package solver.constraints.extension.nary;
 
+import solver.variables.IntVar;
+
 import java.io.Serializable;
 
 public abstract class LargeRelation implements Serializable {
@@ -47,16 +49,11 @@ public abstract class LargeRelation implements Serializable {
      */
     public abstract boolean isConsistent(int[] tuple);
 
-    protected boolean valid(int[] tuple, int[] offsets, int[] domSizes) {
-        /*for (int i = 0; i < tuple.length; i++) {
-                if (!between(tuple[i], offsets[i], offsets[i] + domSizes[i] - 1)) return false;
-            }*/
+    protected boolean valid(int[] tuple, IntVar[] vars) {
+        for (int i = 0; i < tuple.length; i++) {
+            if (!vars[i].contains(tuple[i]))
+                return false;
+        }
         return true;
     }
-
-    private static boolean between(int v, int low, int upp) {
-        return (low <= v) && (v <= upp);
-    }
-
-
 }