+++ choco-solver/src/main/java/solver/constraints/extension/nary/PropLargeGAC2001Positive.java
@@ -110,8 +110,8 @@ public class PropLargeGAC2001Positive extends PropLargeCSP<IterTuplesTable> {
     }
 
     @Override
-    protected IterTuplesTable makeRelation(Tuples tuples, int[] offsets, int[] dsizes) {
-        return new IterTuplesTable(tuples, offsets, dsizes);
+    protected IterTuplesTable makeRelation(Tuples tuples, IntVar[] vars) {
+        return new IterTuplesTable(tuples, vars);
     }
 
     @Override