+++ CHANGES.md
@@ -104,3 +104,4 @@ NEXT MILESTONE
 - Table constraint refactoring (STR2+) (seems to be not idempotent however)
 - Table reformulation of small scalar products
 - Minimum and Maximum over boolean arrays
+- Issue #215: Fix generation of relation based on tuples